requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
twilio==7.3.0
urllib3==1.26.7
Werkzeug==2.0.2
pdftotext==2.2.2
six==1.16.0
camelot-py==0.10.1
ghostscript==0.7
